Major refactor of global optimize
