+++ src/global_optimization/Changes/Change.py
@@ -1,0 +1,14 @@
+useDistanceObjective = False
+
+
+class Change:
+    @property
+    def is_valid(self) -> bool:
+        pass
+
+    @property
+    def costdiff(self) -> float:
+        pass
+
+    def apply(self) -> None:
+        pass
