+++ src/global_optimization/utils.py
@@ -1,0 +1,3 @@
+def gerp(a, b, t):
+    """Geometric interpolation"""
+    return a * (b / a) ** t
