+++ src/main.py
@@ -11,14 +11,18 @@ import argparse
 import multiprocessing
 from pathlib import Path
 import optimization
-import global_optimization
+from global_optimization import global_optimize, auto_temp_schedule
 from scipy.ndimage import distance_transform_edt
-from colony import LineageFrames
 import numpy as np
+from matplotlib import cm
+from matplotlib.colors import Normalize
+from PIL import Image
+from lineage_funcs import create_lineage, save_lineage
 
 import sys
 sys.setrecursionlimit(10000)
 # bootleg fix to prevent recursion error when pickling large lineages
+
 
 def parse_args():
     """Reads and parses the command-line arguments."""
@@ -43,23 +47,22 @@ def parse_args():
                         help='one of "best-wins", "worst-wins", "extreme-wins"')
     parser.add_argument('--cluster', type=str, default='',
                         help='dask cluster address (defaults to local cluster)')
-
     parser.add_argument('--no_parallel', action='store_true', default=False, help='disable parallelism')
     parser.add_argument('--global_optimization', action='store_true', default=False, help='global optimization')
     parser.add_argument('--binary', action='store_true', default=True,
-                        help = "input image is binary")
+                        help="input image is binary")
     parser.add_argument('--graySynthetic', action='store_true', default=False,
                         help='enables the use of the grayscale synthetic image for use with non-thresholded images')
     parser.add_argument('--phaseContrast', action='store_true', default=False,
                         help='enables the use of the grayscale synthetic image for phase contract images')
     parser.add_argument('-ta', '--auto_temp', metavar='TEMP', type=int, default=1,
-                          help='auto temperature scheduling for the simulated annealing')
+                        help='auto temperature scheduling for the simulated annealing')
     parser.add_argument('-ts', '--start_temp', type=float, help='starting temperature for the simulated annealing')
     parser.add_argument('-te', '--end_temp', type=float, help='ending temperature for the simulated annealing')
     parser.add_argument('-am', '--auto_meth', type=str, default='none', choices=('none', 'frame', 'factor', 'const', 'cost'),
                         help='method for auto-temperature scheduling')
     parser.add_argument('-r', "--residual", metavar="FILE", type=Path, required=False,
-                          help="path to the residual image output directory")
+                        help="path to the residual image output directory")
     parser.add_argument('--lineage_file', metavar='FILE', type=Path, required=False,
                         help='path to previous lineage file')
     parser.add_argument('--continue_from', metavar='N', type=int, default=0,
@@ -68,7 +71,6 @@ def parse_args():
     parser.add_argument('--batches', metavar='N', type=int, default=1, help='number of batches to split each frame into for multithreading')
 
     # required arguments
-
     required = parser.add_argument_group('required arguments')
     required.add_argument('-i', '--input', metavar='PATTERN', type=str, required=True,
                           help='input filename pattern (e.g. "image%%03d.png")')
@@ -80,8 +82,6 @@ def parse_args():
                           help='path to the initial cell configuration')
     required.add_argument('-b', "--bestfit", metavar="FILE", type=Path, required=True,
                           help="path to the best fit synthetic image output directory")
-    
-
 
     parsed = parser.parse_args()
 
@@ -121,29 +121,6 @@ def load_config(config_file):
     return config
 
 
-def load_colony(colony, initial_file, config, initial_frame = None):
-    """Loads the initial colony of cells."""
-    with open(initial_file, newline='') as fp:
-        reader = csv.DictReader(fp, skipinitialspace=True)
-        for row in reader:
-            if initial_frame != None and row['file'] != initial_frame:
-                continue
-            name = row['name']
-            celltype = config['global.cellType'].lower()
-            if celltype == 'bacilli':
-                x = float(row['x'])
-                y = float(row['y'])
-                width = float(row['width'])
-                length = float(row['length'])
-                rotation = float(row['rotation'])
-                if config["simulation"]["image.type"] == "graySynthetic":
-                    opacity = config["simulation"]["cell.opacity"]
-                    cell = Bacilli(name, x, y, width, length, rotation, opacity = opacity)
-                else:
-                    cell = Bacilli(name, x, y, width, length, rotation)
-            colony.add(optimization.CellNode(cell))
-
-
 def get_inputfiles(args):
     """Gets the list of images that are to be analyzed."""
     inputfiles = []
@@ -164,8 +141,30 @@ def get_inputfiles(args):
             break
         else:
             raise ValueError(f'Input file not found "{file}"')
-
     return inputfiles
+
+
+def save_output(image_name, synthimage, realimage, cellnodes, args, config):
+    residual_vmin = config["residual.vmin"]
+    residual_vmax = config["residual.vmax"]
+    if args.residual:
+        colormap = cm.ScalarMappable(norm=Normalize(vmin=residual_vmin, vmax=residual_vmax), cmap="bwr")
+    bestfit_frame = Image.fromarray(np.uint8(255 * synthimage), "L")
+    bestfit_frame.save(args.bestfit / image_name)
+    shape = realimage.shape
+    output_frame = np.empty((shape[0], shape[1], 3))
+    output_frame[..., 0] = realimage
+    output_frame[..., 1] = output_frame[..., 0]
+    output_frame[..., 2] = output_frame[..., 0]
+    for node in cellnodes:
+        node.cell.drawoutline(output_frame, (1, 0, 0))
+    output_frame = Image.fromarray(np.uint8(255 * output_frame))
+    output_frame.save(args.output / image_name)
+
+    if args.residual:
+        residual_frame = Image.fromarray(np.uint8(255 * colormap.to_rgba(np.clip(realimage - synthimage,
+                                                                                 residual_vmin, residual_vmax))), "RGB")
+        residual_frame.save(args.residual / image_name)
 
 
 def main(args):
@@ -178,22 +177,13 @@ def main(args):
         from dask.distributed import Client, LocalCluster
         if not args.cluster:
             cluster = LocalCluster(
-                n_workers = args.workers, threads_per_worker = 1,
+                n_workers=args.workers, threads_per_worker=1,
             )
             client = Client(cluster)
         else:
             cluster = args.cluster
             client = Client(cluster)
             client.restart()
-            cwd = Path(__file__).parent.absolute()
-            client.upload_file(cwd / 'drawing.py')
-            client.upload_file(cwd / 'mathhelper.py')
-            client.upload_file(cwd / 'cell.py')
-            client.upload_file(cwd / 'colony.py')
-            client.upload_file(cwd / 'optimization.py')
-            client.upload_file(cwd / 'drawing.py')
-            client.upload_file(cwd / 'global_optimization.py')
-            client.upload_file(cwd / 'main.py')
     else:
         client = None
 
@@ -202,67 +192,55 @@ def main(args):
 
     try:
         config = load_config(args.config)
-        
+
         simulation_config = config["simulation"]
-        #Maybe better to store the image type in the config file in the first place, instead of using cmd?
-        if args.graySynthetic == True:
+        if args.graySynthetic:
             simulation_config["image.type"] = "graySynthetic"
-        elif args.phaseContrast == True:
+        elif args.phaseContrast:
             simulation_config["image.type"] = "phaseContrastImage"
-        elif args.binary == True:
+        elif args.binary:
             simulation_config["image.type"] = "binary"
         else:
             raise ValueError("Invalid Command: Synthetic image type must be specified")
-        
+
         if not args.output.is_dir():
             args.output.mkdir()
         if not args.bestfit.is_dir():
             args.bestfit.mkdir()
         if args.residual and not args.residual.is_dir():
             args.residual.mkdir()
-            
+
         seed = int(start * 1000) % (2**32)
-        if args.seed != None:
+        if args.seed is not None:
             seed = args.seed
         np.random.seed(seed)
         print("Seed: {}".format(seed))
-        
-        
+
         celltype = config['global.cellType'].lower()
 
-        # setup the colony from a file with the initial properties
-        lineageframes = LineageFrames()
-        colony = lineageframes.forward()
-        imagefiles =  get_inputfiles(args)
-        if args.lineage_file:
-            load_colony(colony, args.lineage_file, config, initial_frame = imagefiles[0].name)
-        else:
-            load_colony(colony, args.initial, config)
-        cost_diff = (-1, -1)
+        imagefiles = get_inputfiles(args)
 
         # open the lineage file for writing
-        lineagefile = open(args.output/'lineage.csv', 'w')
+        lineagefile = open(args.output / 'lineage.csv', 'w')
         header = ['file', 'name']
         if celltype == 'bacilli':
             header.extend(['x', 'y', 'width', 'length', 'rotation', "split_alpha", "opacity"])
         print(','.join(header), file=lineagefile)
 
         if args.debug:
-            with open(args.debug/'debug.csv', 'w') as debugfile:
+            with open(args.debug / 'debug.csv', 'w') as debugfile:
                 print(','.join(['window_start', 'window_end', 'pbad_total', 'bad_count', 'temperature', 'total_cost_diff', 'current_iteration', 'total_iterations']), file=debugfile)
-
 
         if args.global_optimization:
             global useDistanceObjective
-            
+
             useDistanceObjective = args.dist
             realimages = [optimization.load_image(imagefile) for imagefile in imagefiles]
+
+            # setup the colony from a file with the initial properties
+            lineage = create_lineage(imagefiles, realimages, config, args)
+
             window = config["global_optimizer.window_size"]
-            if args.lineage_file:
-                lineage = global_optimization.build_initial_lineage(imagefiles, args.lineage_file, args.continue_from, config["simulation"])
-            else:
-                lineage = global_optimization.build_initial_lineage(imagefiles, args.initial, args.continue_from, config["simulation"])
-            lineage = global_optimization.find_optimal_simulation_confs(imagefiles, lineage, realimages, args.continue_from)
             sim_start = args.continue_from - args.frame_first
             print(sim_start)
             shape = realimages[0].shape
@@ -273,10 +251,10 @@ def main(args):
             if not useDistanceObjective:
                 distmaps = [None] * len(realimages)
             for window_start in range(1 - window, len(realimages)):
-                 window_end = window_start + window
-                 print(window_start, window_end)
-                 if window_end <= len(realimages):
-                        # get initial estimate
+                window_end = window_start + window
+                print(window_start, window_end)
+                if window_end <= len(realimages):
+                    # get initial estimate
                     if window_start >= sim_start:
                         if window_end > 1:
                             lineage.copy_forward()
@@ -293,87 +271,28 @@ def main(args):
                     if args.auto_temp == 1 and window_end == 1:
                         print("auto temperature schedule started")
                         args.start_temp, args.end_temp = \
-                            global_optimization.auto_temp_schedule(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, 0, 1, lineagefile, args, config)
+                            auto_temp_schedule(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, 0, 1, lineagefile, args, config)
                         print("auto temperature schedule finished")
                         print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
                     if args.auto_meth == "frame" and optimization.auto_temp_schedule_frame(window_end, 3):
                         print("auto temperature schedule restarted")
                         args.start_temp, args.end_temp = \
-                            global_optimization.auto_temp_schedule(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, window_start, window_end, lineagefile, args, config)
+                            auto_temp_schedule(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, window_start, window_end, lineagefile, args, config)
                         print("auto temperature schedule finished")
                         print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
-                 if window_start >= sim_start:
+                if window_start >= sim_start:
                     if useDistanceObjective:
-                        global_optimization.totalCostDiff = optimization.dist_objective(realimage, synthimage, distmap, cellmap, config["overlap.cost"])
+                        global_optimize.totalCostDiff = optimization.dist_objective(realimage, synthimage, distmap, cellmap, config["overlap.cost"])
                     else:
-                        global_optimization.totalCostDiff = optimization.objective(realimage, synthimage, cellmap, config["overlap.cost"], config["cell.importance"])
-                    lineage, synthimages, distmaps, cellmaps = global_optimization.optimize(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, window_start, window_end, lineagefile, args, config, iteration_per_cell, client = client)
-                 if window_start >= 0:
-                    global_optimization.save_lineage(imagefiles[window_start].name, lineage.frames[window_start].nodes, lineagefile)
-                    global_optimization.save_output(imagefiles[window_start].name, synthimages[window_start], realimages[window_start], lineage.frames[window_start].nodes, args, config)
+                        global_optimize.totalCostDiff = optimization.objective(realimage, synthimage, cellmap, config["overlap.cost"], config["cell.importance"])
+                    lineage, synthimages, distmaps, cellmaps = global_optimize(imagefiles, lineage, realimages, synthimages, cellmaps, distmaps, window_start, window_end, lineagefile, args, config, iteration_per_cell, client=client)
+                if window_start >= 0:
+                    save_lineage(imagefiles[window_start].name, lineage.frames[window_start].nodes, lineagefile)
+                    save_output(imagefiles[window_start].name, synthimages[window_start], realimages[window_start], lineage.frames[window_start].nodes, args, config)
             return 0
-        
-        config["simulation"] = optimization.find_optimal_simulation_conf(config["simulation"], optimization.load_image(imagefiles[0]), list(colony))
-        if args.auto_temp == 1:
-            print("auto temperature schedule started")
-            args.start_temp, args.end_temp = optimization.auto_temp_schedule(imagefiles[0], lineageframes.forward(), args, config)
-            print("auto temperature schedule finished")
-            print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
 
-        frame_num = 0
-        prev_cell_num = len(colony)
-        for imagefile in imagefiles: # Recomputing temperature when needed
-
-            frame_num += 1
-
-            if args.auto_meth == "frame":
-                if optimization.auto_temp_schedule_frame(frame_num, 8):
-                    print("auto temperature schedule started (recomputed)")
-                    args.start_temp, args.end_temp = optimization.auto_temp_schedule(imagefile, colony, args, config)
-                    print("auto temperature schedule finished")
-                    print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
-
-            elif args.auto_meth == "factor":
-                if optimization.auto_temp_schedule_factor(len(colony), prev_cell_num, 1.1):
-                    print("auto temperature schedule started (recomputed)")
-                    args.start_temp, args.end_temp = optimization.auto_temp_schedule(imagefile, colony, args, config)
-                    print("auto temperature schedule finished")
-                    print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
-                    prev_cell_num = len(colony)
-
-            elif args.auto_meth == "const":
-                if optimization.auto_temp_schedule_const(len(colony), prev_cell_num, 10):
-                    print("auto temperature schedule started (recomputed)")
-                    args.start_temp, args.end_temp = optimization.auto_temp_schedule(imagefile, colony, args, config)
-                    print("auto temperature schedule finished")
-                    print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
-                    prev_cell_num = len(colony)
-
-            elif args.auto_meth == "cost":
-                print(cost_diff, frame_num, optimization.auto_temp_shcedule_cost(cost_diff))
-                if frame_num >= 2 and optimization.auto_temp_shcedule_cost(cost_diff):
-                    print("auto temperature schedule started cost_diff (recomputed)")
-                    args.start_temp, args.end_temp = optimization.auto_temp_schedule(imagefile, colony, args, config)
-                    print("auto temperature schedule finished")
-                    print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
-
-            colony = optimize(imagefile, lineageframes, args, config, client)
-
-            cost_diff = optimization.update_cost_diff(colony, cost_diff)
-
-            # flatten modifications and save cell properties
-
-            colony.flatten()
-            for cellnode in colony:
-                properties = [imagefile.name, cellnode.cell.name]
-                if celltype == 'bacilli':                                   
-                    properties.extend([
-                        str(cellnode.cell.x),
-                        str(cellnode.cell.y),
-                        str(cellnode.cell.width),
-                        str(cellnode.cell.length),
-                        str(cellnode.cell.rotation)])
-                print(','.join(properties), file=lineagefile)
+        # local optimization
+        optimization.local_optimize(imagefiles, config, args, lineagefile, client)
 
     except KeyboardInterrupt as error:
         raise error
@@ -397,7 +316,6 @@ if __name__ == '__main__':
 
     import jsonc
     from cell import Bacilli
-    from optimization import optimize
     from sys import exit
     # pr = cProfile.Profile()
     # pr.enable()
