+++ src/optimization.py
@@ -9,7 +9,8 @@ from scipy.ndimage import distance_transform_edt
 from scipy.optimize import leastsq
 
 from cell import Bacilli
-from colony import CellNode, Colony
+from colony import LineageFrames
+from lineage_funcs import load_colony
 
 import dask.distributed
 
@@ -21,25 +22,27 @@ is_cell = True
 is_background = False
 
 
-
 def objective(realimage, synthimage, cellmap, overlap_cost, cell_importance):
     """Full objective function between two images."""
-    overlap_map = cellmap[cellmap>1] - 1
+    overlap_map = cellmap[cellmap > 1] - 1
     return np.sum(np.square((realimage - synthimage))) \
-        + overlap_cost * np.sum(np.square(overlap_map)) 
+        + overlap_cost * np.sum(np.square(overlap_map))
+
 
 def dist_objective(realimage, synthimage, distmap, cellmap, overlap_cost):
-    overlap_map = cellmap[cellmap>1] - 1
-    return np.sum(np.square((realimage-synthimage)*distmap)) + overlap_cost * np.sum(np.square(overlap_map))
+    overlap_map = cellmap[cellmap > 1] - 1
+    return np.sum(np.square((realimage - synthimage) * distmap)) + overlap_cost * np.sum(np.square(overlap_map))
+
 
 def find_optimal_simulation_conf(simulation_config, realimage1, cellnodes):
     shape = realimage1.shape
+
     def cost(values, target, simulation_config):
         for i in range(len(target)):
             simulation_config[target[i]] = values[i]
         synthimage, cellmap = generate_synthetic_image(cellnodes, shape, simulation_config)
         return (realimage1 - synthimage).flatten()
-    
+
     initial_values = []
     variables = []
     if simulation_config["background.color"] == "auto":
@@ -61,26 +64,26 @@ def find_optimal_simulation_conf(simulation_config, realimage1, cellnodes):
     if len(variables) != 0:
         residues = lambda x: cost(x, variables, simulation_config)
         optimal_values, _ = leastsq(residues, initial_values)
-        
+
         for i, param in enumerate(variables):
             simulation_config[param] = optimal_values[i]
         simulation_config["cell.opacity"] = max(0, simulation_config["cell.opacity"])
         simulation_config["light.diffraction.sigma"] = max(0, simulation_config["light.diffraction.sigma"])
-        
+
         if auto_opacity:
             for node in cellnodes:
                 node.cell.opacity = simulation_config["cell.opacity"]
-            
+
     print(f"optimal simulation configuration values found: {simulation_config}")
     return simulation_config
-            
+
 
 def generate_synthetic_image(cellnodes, shape, simulation_config):
     image_type = simulation_config["image.type"]
     cellmap = np.zeros(shape, dtype=int)
     if image_type == "graySynthetic" or image_type == "phaseContrast":
         background_color = simulation_config["background.color"]
-        synthimage = np.full(shape, background_color) 
+        synthimage = np.full(shape, background_color)
         for node in cellnodes:
             node.cell.draw(synthimage, cellmap, is_cell, simulation_config)
         return synthimage, cellmap
@@ -91,19 +94,18 @@ def generate_synthetic_image(cellnodes, shape, simulation_config):
         return synthimage, cellmap
 
 
-
 def load_image(imagefile):
     """Open the image file and convert to a floating-point grayscale array."""
     with open(imagefile, 'rb') as fp:
         realimage = np.array(Image.open(fp))
     if realimage.dtype == np.uint8:
-        realimage = realimage.astype(float)/255
+        realimage = realimage.astype(float) / 255
     if len(realimage.shape) == 3:
         realimage = np.mean(realimage, axis=-1)
     return realimage
 
 
-def perturb_bacilli(node, config, imageshape, invalid_limit = 50):
+def perturb_bacilli(node, config, imageshape, invalid_limit=50):
     """Create a new perturbed bacilli cell."""
     global badcount  # DEBUG
     cell = node.cell
@@ -118,41 +120,39 @@ def perturb_bacilli(node, config, imageshape, invalid_limit = 50):
         else:
             AssertionError('Names not matching')
 
-    max_displacement = config['bacilli.maxSpeed']/config['global.framesPerSecond']
-    max_rotation = config['bacilli.maxSpin']/config['global.framesPerSecond']
+    max_displacement = config['bacilli.maxSpeed'] / config['global.framesPerSecond']
+    max_rotation = config['bacilli.maxSpin'] / config['global.framesPerSecond']
     min_growth = config['bacilli.minGrowth']
     max_growth = config['bacilli.maxGrowth']
     min_width = config['bacilli.minWidth']
     max_width = config['bacilli.maxWidth']
     min_length = config['bacilli.minLength']
     max_length = config['bacilli.maxLength']
-    
+
     perturb_conf = config["perturbation"]
     p_x = perturb_conf["prob.x"]
     p_y = perturb_conf["prob.y"]
     p_width = perturb_conf["prob.width"]
     p_length = perturb_conf["prob.length"]
     p_rotation = perturb_conf["prob.rotation"]
-    
+
     x_mu = perturb_conf["modification.x.mu"]
     y_mu = perturb_conf["modification.y.mu"]
     width_mu = perturb_conf["modification.width.mu"]
     length_mu = perturb_conf["modification.length.mu"]
     rotation_mu = perturb_conf["modification.rotation.mu"]
-    
+
     x_sigma = perturb_conf["modification.x.sigma"]
     y_sigma = perturb_conf["modification.y.sigma"]
     width_sigma = perturb_conf["modification.width.sigma"]
     length_sigma = perturb_conf["modification.length.sigma"]
     rotation_sigma = perturb_conf["modification.rotation.sigma"]
-    
+
     invalid_count = 0
     simulation_config = config["simulation"]
     if simulation_config["image.type"] == "graySynthetic":
         p_opacity = perturb_conf["prob.opacity"]
-        opacity_mu = perturb_conf["modification.opacity.mu"]
-        opacity_sigma = perturb_conf["modification.opacity.sigma"]
-    
+
     while invalid_count < invalid_limit:
         # set starting properties
         x = cell.x
@@ -161,55 +161,57 @@ def perturb_bacilli(node, config, imageshape, invalid_limit = 50):
         length = cell.length
         rotation = cell.rotation
         cell_opacity = cell.opacity
-        
+
         if simulation_config["image.type"] == "graySynthetic":
             p_decision = np.array([p_x, p_y, p_width, p_length, p_rotation, p_opacity])
         else:
             p_decision = np.array([p_x, p_y, p_width, p_length, p_rotation])
-            
-        p = np.random.uniform(0.0, 1.0, size= p_decision.size)
-        
+
+        p = np.random.uniform(0.0, 1.0, size=p_decision.size)
+
         # generate a sequence such that at least an attribute must be modified
         while (p > p_decision).all():
-            p = np.random.uniform(0.0, 1.0, size= p_decision.size)
-        
-        if p[0] < p_decision[0]: #perturb x
+            p = np.random.uniform(0.0, 1.0, size=p_decision.size)
+
+        if p[0] < p_decision[0]:  # perturb x
             x = cell.x + np.random.normal(x_mu, x_sigma)
-    
-        if p[1] < p_decision[1]: #perturb y
+
+        if p[1] < p_decision[1]:  # perturb y
             y = cell.y + np.random.normal(y_mu, y_sigma)
-    
-        if p[2] < p_decision[2]: #perturb width
+
+        if p[2] < p_decision[2]:  # perturb width
             width = cell.width + np.random.normal(width_mu, width_sigma)
-    
-        if p[3] < p_decision[3]: #perturb length
+
+        if p[3] < p_decision[3]:  # perturb length
             length = cell.length + np.random.normal(length_mu, length_sigma)
-    
-        if p[4] < p_decision[4]: #perturb rotation
+
+        if p[4] < p_decision[4]:  # perturb rotation
             rotation = cell.rotation + np.random.normal(rotation_mu, rotation_sigma)
-        #if simulation_config["image.type"] == "graySynthetic" and p[5] < p_decision[5]:
-            #cell_opacity = cell.opacity + (np.random.normal(opacity_mu, opacity_sigma))
-            
+        # if simulation_config["image.type"] == "graySynthetic" and p[5] < p_decision[5]:
+            # cell_opacity = cell.opacity + (np.random.normal(opacity_mu, opacity_sigma))
+
         displacement = sqrt(np.sum((np.array([x, y, 0] - prior.position))**2))
-        
-        if not (0 <= x < imageshape[1] and 0 <= y < imageshape[0]) or (displacement > max_displacement)\
-            or width < min_width or width > max_width or (abs(rotation - prior.rotation) > max_rotation) or \
-            not (min_length < length < max_length) or not (min_growth < length - prior.length < max_growth):
-                badcount += 1
-                invalid_count+=1
+
+        if not (0 <= x < imageshape[1] and 0 <= y < imageshape[0]) or (displacement > max_displacement) \
+           or width < min_width or width > max_width or (abs(rotation - prior.rotation) > max_rotation) or \
+           not (min_length < length < max_length) or not (min_growth < length - prior.length < max_growth):
+            badcount += 1
+            invalid_count += 1
         elif simulation_config["image.type"] == "graySynthetic" and cell_opacity < 0:
             badcount += 1
-            invalid_count+=1
+            invalid_count += 1
         else:
             break
-        
+
     # push the new cell over the previous in the node
     node.push(Bacilli(cell.name, x, y, width, length, rotation, cell.split_alpha, cell_opacity))
+
 
 def split_proba(length):
     """Returns the split probability given the length of the cell."""
     # Determined empirically based on previous runs
     return min(1, math.sin((length - 14) / (2 * math.pi * math.pi))) if 14 < length else 0
+
 
 def bacilli_split(node, config, imageshape):
     """Split the cell and push both onto the stack for testing."""
@@ -217,14 +219,14 @@ def bacilli_split(node, config, imageshape):
     if np.random.random_sample() > split_proba(node.cell.length):
         return False
 
-    max_displacement = config['bacilli.maxSpeed']/config['global.framesPerSecond']
-    max_rotation = config['bacilli.maxSpin']/config['global.framesPerSecond']
+    max_displacement = config['bacilli.maxSpeed'] / config['global.framesPerSecond']
+    max_rotation = config['bacilli.maxSpin'] / config['global.framesPerSecond']
     min_width = config['bacilli.minWidth']
     max_width = config['bacilli.maxWidth']
     min_length = config['bacilli.minLength']
     max_length = config['bacilli.maxLength']
 
-    alpha = np.random.random_sample()/5 + 2/5     # TODO choose from config
+    alpha = np.random.random_sample() / 5 + 2/5     # TODO choose from config
     cell1, cell2 = node.cell.split(alpha)
 
     # make sure that the lengths are within constraints
@@ -273,8 +275,8 @@ def bacilli_combine(node, config, imageshape):
     if np.random.random_sample() > 0.2:
         return False, None
 
-    max_displacement = config['bacilli.maxSpeed']/config['global.framesPerSecond']
-    max_rotation = config['bacilli.maxSpin']/config['global.framesPerSecond']
+    max_displacement = config['bacilli.maxSpeed'] / config['global.framesPerSecond']
+    max_rotation = config['bacilli.maxSpin'] / config['global.framesPerSecond']
     min_width = config['bacilli.minWidth']
     max_width = config['bacilli.maxWidth']
     min_length = config['bacilli.minLength']
@@ -330,18 +332,23 @@ def bacilli_combine(node, config, imageshape):
 
     return True, presplit
 
+
 # functions for different types of dynamic auto-temperature scheduling
 def auto_temp_schedule_frame(frame, k_frame):
     return frame % k_frame == 0
 
+
 def auto_temp_schedule_factor(cell_num, prev_num, factor):
     return True if cell_num / prev_num >= factor else False
+
 
 def auto_temp_schedule_const(cell_num, prev_num, constant):
     return True if cell_num - prev_num >= constant else False
 
-def auto_temp_shcedule_cost(cost_diff):
+
+def auto_temp_schedule_cost(cost_diff):
     return True if ((cost_diff[1] - cost_diff[0]) / cost_diff[0]) > 0.2 else False
+
 
 def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, auto_temp_complete=True, auto_const_temp = 1):
     """Core of the optimization routine."""
@@ -353,7 +360,7 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
     realimage = load_image(imagefile)
     shape = realimage.shape
     simulation_config = config["simulation"]
-    
+
     celltype = config['global.cellType'].lower()
     useDistanceObjective = args.dist
 
@@ -363,14 +370,14 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
     synthimage, cellmap = generate_synthetic_image(cellnodes, shape, simulation_config)
     if useDistanceObjective:
         distmap = distance_transform_edt(realimage < .5)
-        distmap /= config[f'{celltype}.distanceCostDivisor']*config['global.pixelsPerMicron']
+        distmap /= config[f'{celltype}.distanceCostDivisor'] * config['global.pixelsPerMicron']
         distmap += 1
         cost = dist_objective(realimage, synthimage, distmap, cellmap, config["overlap.cost"])
     else:
         cost = objective(realimage, synthimage, cellmap, config["overlap.cost"], config["cell.importance"])
 
     # setup temperature schedule
-    run_count = int(iterations_per_cell*len(cellnodes))
+    run_count = int(iterations_per_cell * len(cellnodes))
 
     if (auto_temp_complete == False):
         temperature = auto_const_temp
@@ -378,7 +385,7 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
         temperature = args.start_temp
         end_temperature = args.end_temp
 
-        alpha = (end_temperature/temperature)**(1/run_count)
+        alpha = (end_temperature / temperature)**(1 / run_count)
 
     for i in range(run_count):
         # print progress for debugging purposes
@@ -413,8 +420,8 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
                 snode1, snode2 = cnode.children
 
                 # compute the starting cost
-                region = (cnode.cell.simulated_region(simulation_config).\
-                          union(snode1.cell.simulated_region(simulation_config))\
+                region = (cnode.cell.simulated_region(simulation_config).
+                          union(snode1.cell.simulated_region(simulation_config))
                           .union(snode2.cell.simulated_region(simulation_config)))
                 if useDistanceObjective:
                     start_cost = dist_objective(
@@ -457,7 +464,7 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
                                            cellmap[region.top:region.bottom,region.left:region.right],
                                            config["overlap.cost"],
                                            config["cell.importance"])
-                
+
                 # subtract the previous cell
                 node.cell.draw(synthimage, cellmap, is_background, simulation_config)
 
@@ -498,17 +505,17 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
                         config["overlap.cost"])
             else:
                 end_cost = objective(realimage[region.top:region.bottom,region.left:region.right],
-                                       synthimage[region.top:region.bottom,region.left:region.right],
-                                       cellmap[region.top:region.bottom,region.left:region.right],
-                                       config["overlap.cost"],
-                                       config["cell.importance"])
+                                     synthimage[region.top:region.bottom,region.left:region.right],
+                                     cellmap[region.top:region.bottom,region.left:region.right],
+                                     config["overlap.cost"],
+                                     config["cell.importance"])
             costdiff = end_cost - start_cost
 
             # compute the acceptance threshold
             if costdiff <= 0:
                 acceptance = 1.0
             else:
-                acceptance = np.exp(-costdiff/temperature)
+                acceptance = np.exp(-costdiff / temperature)
                 bad_count += 1
                 bad_prob_tot += acceptance
 
@@ -534,49 +541,47 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
             colony.flatten()
             cellnodes = list(colony)
 
-            
             # DEBUG
-            if args.debug and i%80 == 0:
+            if args.debug and i % 80 == 0:
                 synthimage, cellmap = generate_synthetic_image(cellnodes, shape, simulation_config)
                 frame_1 = np.empty((shape[0], shape[1], 3))
                 frame_1[..., 0] = (realimage - synthimage)
                 frame_1[..., 1] = frame_1[..., 0]
                 frame_1[..., 2] = frame_1[..., 0]
 
-                #for node in cellnodes:
-                    #node.cell.drawoutline(frame_1, (1, 0, 0))
+                # for node in cellnodes:
+                    # node.cell.drawoutline(frame_1, (1, 0, 0))
 
                 frame_1 = np.clip(frame_1, 0, 1)
 
-                debugimage = Image.fromarray((255*frame_1).astype(np.uint8))
-                debugimage.save(args.debug/f'residual{debugcount}.png')
-
+                debugimage = Image.fromarray((255 * frame_1).astype(np.uint8))
+                debugimage.save(args.debug / f'residual{debugcount}.png')
 
                 frame_2 = np.empty((shape[0], shape[1], 3))
                 frame_2[..., 0] = synthimage
                 frame_2[..., 1] = frame_2[..., 0]
                 frame_2[..., 2] = frame_2[..., 0]
 
-                #for node in cellnodes:
-                    #node.cell.drawoutline(frame_2, (1, 0, 0))
+                # for node in cellnodes:
+                    #.node.cell.drawoutline(frame_2, (1, 0, 0))
 
                 frame_2 = np.clip(frame_2, 0, 1)
 
-                debugimage = Image.fromarray((255*frame_2).astype(np.uint8))
-                debugimage.save(args.debug/f'synthetic{debugcount}.png')
+                debugimage = Image.fromarray((255 * frame_2).astype(np.uint8))
+                debugimage.save(args.debug / f'synthetic{debugcount}.png')
                 debugcount += 1
 
-        if (auto_temp_complete == True):
+        if auto_temp_complete:
             temperature *= alpha
 
-    #print(f'Bad Percentage: {100*badcount/run_count}%')
+    # print(f'Bad Percentage: {100*badcount/run_count}%')
 
-    if (auto_temp_complete == False):
+    if not auto_temp_complete:
 
-        print("pbad is ", bad_prob_tot/bad_count)
-        #print("temperature is ", temperature)
+        print("pbad is ", bad_prob_tot / bad_count)
+        # print("temperature is ", temperature)
 
-        return bad_prob_tot/bad_count
+        return bad_prob_tot / bad_count
 
     if useDistanceObjective:
         new_cost = dist_objective(realimage, synthimage, distmap, cellmap, config["overlap.cost"])
@@ -600,44 +605,46 @@ def optimize_core(imagefile, colony, args, config, iterations_per_cell=3000, aut
 
     frame = np.clip(frame, 0, 1)
 
-    debugimage = Image.fromarray((255*frame).astype(np.uint8))
+    debugimage = Image.fromarray((255 * frame).astype(np.uint8))
 
     best_fit_frame = np.empty((shape[0], shape[1], 3))
     best_fit_frame[..., 0] = synthimage
-    best_fit_frame[..., 1] = best_fit_frame[...,0]
-    best_fit_frame[..., 2] = best_fit_frame[...,0]
+    best_fit_frame[..., 1] = best_fit_frame[..., 0]
+    best_fit_frame[..., 2] = best_fit_frame[..., 0]
     best_fit_frame = np.clip(best_fit_frame, 0, 1)
-    best_fit_frame = Image.fromarray((255* best_fit_frame). astype(np.uint8))
+    best_fit_frame = Image.fromarray((255 * best_fit_frame). astype(np.uint8))
     return colony, cost, debugimage, best_fit_frame
+
 
 def auto_temp_schedule(imagefile, colony, args, config):
     initial_temp = 1
     ITERATION = 500
     AUTO_TEMP_COMPLETE = False
-    
+
     count = 0
 
-    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp)<0.40):
+    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp) < 0.40):
         count += 1
         initial_temp *= 10.0
         print(f"count: {count}")
     print("finished < 0.4")
-    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp)>0.40):
+    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp) > 0.40):
         count += 1 
         initial_temp /= 10.0
         print(f"count: {count}")
     print("finished > 0.4")
-    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp)<0.40):
+    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, initial_temp) < 0.40):
         count += 1
         initial_temp *= 1.1
         print(f"count: {count}")
     end_temp = initial_temp
     print("finished < 0.4")
-    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, end_temp)>=1e-10):
+    while(optimize_core(imagefile, colony, args, config, ITERATION, AUTO_TEMP_COMPLETE, end_temp) >= 1e-10):
         count += 1
         end_temp /= 10.0
 
     return initial_temp, end_temp
+
 
 def optimize(imagefile, lineageframes, args, config, client):
     """Optimize the cell properties using simulated annealing."""
@@ -645,12 +652,12 @@ def optimize(imagefile, lineageframes, args, config, client):
     badcount = 0  # DEBUG
 
     if not client:
-        colony,_,debugimage, best_fit_frame = optimize_core(imagefile, lineageframes.forward(), args, config)
+        colony, _, debugimage, best_fit_frame = optimize_core(imagefile, lineageframes.forward(), args, config)
         debugimage.save(args.output / imagefile.name)
-        best_fit_frame.save(args.bestfit/imagefile.name)
+        best_fit_frame.save(args.bestfit / imagefile.name)
         return colony
 
-    #tasks = []
+    # tasks = []
 
     group = lineageframes.latest_group
     ejob = args.jobs // len(group)
@@ -697,9 +704,10 @@ def optimize(imagefile, lineageframes, args, config, client):
 
     for i, s in enumerate(winning):
         debugimage = s[2]
-        debugimage.save(args.output/'{:03d}-{}'.format(i, imagefile.name))
+        debugimage.save(args.output / '{:03d}-{}'.format(i, imagefile.name))
 
     return colony
+
 
 def update_cost_diff(colony, cost_diff):
     if -1 == cost_diff[0]:
@@ -710,3 +718,75 @@ def update_cost_diff(colony, cost_diff):
         cost_diff = (cost_diff[1], colony.cost)
 
     return cost_diff
+
+
+def local_optimize(imagefiles, config, args, lineagefile, client):
+    lineageframes = LineageFrames()
+    colony = lineageframes.forward()
+    if args.lineage_file:
+        load_colony(colony, args.lineage_file, config, initial_frame=imagefiles[0].name)
+    else:
+        load_colony(colony, args.initial, config)
+    cost_diff = (-1, -1)
+    celltype = config['global.cellType'].lower()
+
+    config["simulation"] = find_optimal_simulation_conf(config["simulation"], load_image(imagefiles[0]), list(colony))
+    if args.auto_temp == 1:
+        print("auto temperature schedule started")
+        args.start_temp, args.end_temp = auto_temp_schedule(imagefiles[0], lineageframes.forward(), args, config)
+        print("auto temperature schedule finished")
+        print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
+
+    frame_num = 0
+    prev_cell_num = len(colony)
+    for imagefile in imagefiles:  # Recomputing temperature when needed
+
+        frame_num += 1
+
+        if args.auto_meth == "frame":
+            if auto_temp_schedule_frame(frame_num, 8):
+                print("auto temperature schedule started (recomputed)")
+                args.start_temp, args.end_temp = auto_temp_schedule(imagefile, colony, args, config)
+                print("auto temperature schedule finished")
+                print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
+
+        elif args.auto_meth == "factor":
+            if auto_temp_schedule_factor(len(colony), prev_cell_num, 1.1):
+                print("auto temperature schedule started (recomputed)")
+                args.start_temp, args.end_temp = auto_temp_schedule(imagefile, colony, args, config)
+                print("auto temperature schedule finished")
+                print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
+                prev_cell_num = len(colony)
+
+        elif args.auto_meth == "const":
+            if auto_temp_schedule_const(len(colony), prev_cell_num, 10):
+                print("auto temperature schedule started (recomputed)")
+                args.start_temp, args.end_temp = auto_temp_schedule(imagefile, colony, args, config)
+                print("auto temperature schedule finished")
+                print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
+                prev_cell_num = len(colony)
+
+        elif args.auto_meth == "cost":
+            print(cost_diff, frame_num, auto_temp_schedule_cost(cost_diff))
+            if frame_num >= 2 and auto_temp_schedule_cost(cost_diff):
+                print("auto temperature schedule started cost_diff (recomputed)")
+                args.start_temp, args.end_temp = auto_temp_schedule(imagefile, colony, args, config)
+                print("auto temperature schedule finished")
+                print("starting temperature is ", args.start_temp, "ending temperature is ", args.end_temp)
+
+        colony = optimize(imagefile, lineageframes, args, config, client)
+
+        cost_diff = update_cost_diff(colony, cost_diff)
+
+        colony.flatten()
+
+        for cellnode in colony:
+            properties = [imagefile.name, cellnode.cell.name]
+            if celltype == 'bacilli':
+                properties.extend([
+                    str(cellnode.cell.x),
+                    str(cellnode.cell.y),
+                    str(cellnode.cell.width),
+                    str(cellnode.cell.length),
+                    str(cellnode.cell.rotation)])
+            print(','.join(properties), file=lineagefile)
